Move unit tests for parse_size to the correct place
